requirements.txt
Flask
Flask-SQLAlchemy
sqlalchemy-migrate
flask-marshmallow
marshmallow-sqlalchemy
beautifulsoup4
pytest
Faker
flask_wtf
pytz
flask_security